+++ src/main/java/hu/belrol/archeobase/meret/MeretMapper.java
@@ -11,7 +11,7 @@ public class MeretMapper {
         while (rs.next()) {
             Meret meret = new Meret();
             meret.setId(rs.getInt(rs.findColumn("id")));
-            meret.setMeret(rs.getDouble(rs.findColumn("meret")));
+            meret.setMeret(rs.getInt(rs.findColumn("meret")));
             meret.setMeretTipus(rs.getString(rs.findColumn("meret_tipus")));
             meret.setMertekegyseg(Mertekegyseg.valueOf(rs.getString(rs.findColumn("mertekegyseg"))));
             meret.setStratId(rs.getInt(rs.findColumn("strat_id")));
